consolidate types into one module (#3)
