+++ kerngen/pisa_generators/basic.py
@@ -8,9 +8,7 @@ from typing import ClassVar
 
 import high_parser.pisa_operations as pisa_op
 from high_parser.pisa_operations import PIsaOp
-from high_parser.parser import Context, Immediate
-from high_parser.highop import HighOp, expand_ios
-from high_parser.polys import Polys
+from high_parser import Context, Immediate, HighOp, expand_ios, Polys
 
 
 @dataclass
